Convert Achievement tabs to dropdown + search
Instead of using the bloated tabs, the system will use a dropdown to
filter the achievements into a single table, which will additionally
allow the "all" option to be loaded.

An extra "search" textbox has also been added to allow filtering of the
current achievements (of the category) even further.